fixed casting issues to account for arm issues

+++ internal/system/system_linux.go
@@ -22,21 +22,18 @@ func GetSystemInfo() (SystemInfo, error) {
 	si := SystemInfo{
 		"Hostname": hostname,
 		"Uptime":   formatDuration((time.Second * time.Duration(t.Uptime)).Round(time.Second)),
-		"Memory":   fmt.Sprintf("%s / %s", byteSize(t.Freeram), byteSize(uint64(t.Totalram))),
+		"Memory":   fmt.Sprintf("%s / %s", byteSize(uint64(t.Freeram)), byteSize(uint64(t.Totalram))),
 	}
-
 	if lanStats, err := getInterfaceStatsByName(config.Config.DHCP.Interface); err == nil {
 		si["LAN Interface"] = config.Config.DHCP.Interface
 		si["LAN Tx"] = byteSize(uint64(lanStats.TxBytes))
 		si["LAN Rx"] = byteSize(uint64(lanStats.RxBytes))
 	}
-
 	if lanStats, err := getInterfaceStatsByName("ppp0"); err == nil {
 		si["WAN Interface"] = "ppp0"
 		si["WAN Tx"] = byteSize(uint64(lanStats.TxBytes))
 		si["WAN Rx"] = byteSize(uint64(lanStats.RxBytes))
 	}
-
 	return si, nil
 }
 
@@ -53,7 +50,6 @@ const (
 func byteSize(bytes uint64) string {
 	unit := ""
 	value := float64(bytes)
-
 	switch {
 	case bytes >= EXABYTE:
 		unit = "E"
@@ -78,51 +74,41 @@ func byteSize(bytes uint64) string {
 	case bytes == 0:
 		return "0B"
 	}
-
 	result := strconv.FormatFloat(value, 'f', 1, 32)
 	result = strings.TrimSuffix(result, ".0")
 	return result + unit
 }
-
 func getInterfaceStatsByName(interfaceName string) (*InterfaceStatistics, error) {
 	txdat, err := ioutil.ReadFile(fmt.Sprintf("/sys/class/net/%s/statistics/tx_bytes", interfaceName))
 	if err != nil {
 		return nil, err
 	}
-
 	txBytes, err := strconv.ParseUint(strings.TrimRight(string(txdat), "\n"), 10, 64)
 	if err != nil {
 		return nil, err
 	}
-
 	rxdat, err := ioutil.ReadFile(fmt.Sprintf("/sys/class/net/%s/statistics/rx_bytes", interfaceName))
 	if err != nil {
 		return nil, err
 	}
-
 	rxBytes, err := strconv.ParseUint(strings.TrimRight(string(rxdat), "\n"), 10, 64)
 	if err != nil {
 		return nil, err
 	}
-
 	return &InterfaceStatistics{
 		TxBytes: txBytes,
 		RxBytes: rxBytes,
 	}, nil
 }
-
 func formatDuration(d time.Duration) string {
 	if d == 0 {
 		return "0s"
 	}
-
 	days := int(d.Hours()) / 24
 	hours := int(d.Hours()) % 24
 	minutes := int(d.Minutes()) % 60
 	seconds := int(d.Seconds()) % 60
-
 	var parts []string
-
 	if days > 0 {
 		parts = append(parts, fmt.Sprintf("%dd", days))
 	}
@@ -135,6 +121,5 @@ func formatDuration(d time.Duration) string {
 	if seconds > 0 {
 		parts = append(parts, fmt.Sprintf("%ds", seconds))
 	}
-
 	return strings.Join(parts, ":")
 }
